Move atomic swap component into own file.
This is a preparation for adding an additional feature: lending and borrowing.
